feat(metaflag): There was an issue with metadata.
The metadata length was always included in every payload and the
metadata flag was never set in the metadata flags.  Now the setupFrame
and the errorFrame correctly set the flag and the encodePayload
correctly puts the metadata length only when needed.

Second, I refactored the tests so that the test data file does not
become an endless code duplication for all the different cases.  I also
made one function for test data generation which will encode the
payload and set the metadata flag.  I initially was resistant to this
idea but the amount of times I messed it up, when coding by hand, was
to high and I only went through two frames.  Best to simply generate
that piece of data.