feat(respond): The server now has a way to respond to ReqRes
The server is now able to respond to request-responses.  The 'on'
function, when reqres type is passed in, will respond with the frame,
`req`, and a SingleResponse object, `res`.  The conventions look very
similar to other routers, but are specific to RS.  The client should be
ready to perform some operations against other RS servers. Further
tests would be to connect several clients to the server and attemp to
stress test the server.